# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Master API
  description: |
    The Task Master API provides comprehensive project and task management capabilities with multi-tenant organization support.

    ## Features
    - Multi-tenant architecture with organization management
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Comprehensive task and project management
    - Real-time capabilities with webhooks
    - Advanced filtering and search

    ## Authentication
    All endpoints require authentication using Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```

    ## Rate Limiting
    - Authentication endpoints: 10 requests per minute per IP
    - General API endpoints: 1000 requests per hour per user
    - Bulk operations: 100 requests per hour per user
  version: 1.0.0
  contact:
    name: Task Master API Support
    email: api-support@taskmaster.com
    url: https://docs.taskmaster.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.taskmaster.com/v1
    description: Production server
  - url: http://localhost:3001/api/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authentication
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Invalid input data'
            details:
              type: array
              items:
                type: object
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100

    # User and authentication schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        email:
          type: string
          format: email
          example: 'john.doe@example.com'
        fullName:
          type: string
          example: 'John Doe'
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: 'https://example.com/avatar.jpg'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-10T00:00:00Z'
    Tokens:
      type: object
      properties:
        accessToken:
          type: string
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
        refreshToken:
          type: string
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
        expiresIn:
          type: integer
          example: 3600

    # Organization schemas
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '650e8400-e29b-41d4-a716-446655440001'
        name:
          type: string
          example: 'Acme Corporation'
        description:
          type: string
          nullable: true
          example: 'A sample organization'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
    OrganizationWithMembership:
      allOf:
        - $ref: '#/components/schemas/Organization'
        - type: object
          properties:
            role:
              type: string
              enum: [admin, member]
              example: 'admin'
            memberCount:
              type: integer
              example: 5
            projectCount:
              type: integer
              example: 3
            joinedAt:
              type: string
              format: date-time
              example: '2024-01-01T00:00:00Z'
...
    SignupRequest:
      type: object
      required:
        - fullName
        - email
        - password
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 100
          example: 'John Doe'
        email:
          type: string
          format: email
          example: 'john.doe@example.com'
        password:
          type: string
          minLength: 8
          example: 'SecurePassword123!'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: 'john.doe@example.com'
        password:
          type: string
          example: 'SecurePassword123!'
...
    CreateOrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          example: 'Acme Corporation'
        description:
          type: string
          maxLength: 500
          example: 'A sample organization'
...
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
...
paths:
...
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Sign up a new user
      description: Create a new user account. Email verification is required before the user can sign in.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: 'Registration successful. Please check your email to verify your account.'
                          user:
                            type: object
                            properties:
                              id:
                                type: string
                                format: uuid
                              email:
                                type: string
                                format: email
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Sign in user
      description: Authenticate user and return access and refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
                          tokens:
                            $ref: '#/components/schemas/Tokens'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send a password reset email to the user.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: 'john.doe@example.com'
      responses:
        '200':
          description: Password reset email sent (if account exists)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: 'If an account exists with this email, a password reset link has been sent.'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset user password using the token from email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                  example: 'reset-token-from-email'
                newPassword:
                  type: string
                  minLength: 8
                  example: 'NewSecurePassword123!'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: 'Password has been reset successfully'

  # User profile endpoints
...
  /organizations:
    get:
      tags:
        - Organizations
      summary: List organizations
      description: Get all organizations for the current user.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          organizations:
                            type: array
                            items:
                              $ref: '#/components/schemas/OrganizationWithMembership'

    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          organization:
                            $ref: '#/components/schemas/Organization'
                          membership:
                            type: object
                            properties:
                              role:
                                type: string
                                example: 'admin'
                              joinedAt:
                                type: string
                                format: date-time
...